backend/pyproject.toml
[project]
name = "ipg"
version = "0.1.0"
description = "Islamic Party Game - Real-time multiplayer platform for Islamicized party games"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "passlib>=1.7.4",
    "loguru>=0.7.2",
    "scalar-fastapi>=1.0.0",
    "httpx>=0.27.0",
    "pycountry>=23.12.11",
    "email-validator>=2.2.0",
    "python-multipart>=0.0.12",
    "slowapi>=0.1.9",
    "uvloop>=0.21.0",
    "httptools>=0.6.0",
    "redis[hiredis]>=5.0.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "resend>=2.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "faker>=30.0.0",
    "freezegun>=1.4.0",
    "testcontainers>=4.8.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "types-passlib>=1.7.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "C90", "N", "PL", "ARG"]
ignore = ["E501", "PLR0913", "PLR2004", "ARG002", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["ipg"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
addopts = "--cov=ipg --cov-report=term --cov-fail-under=50"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
omit = [
    "ipg/app.py",
    "ipg/dependencies.py",
    "ipg/database.py",
    "ipg/logger_config.py",
    "ipg/settings.py",
    "scripts/*",
]

[tool.poe.tasks]
# Check commands (CI mode)
lint = "ruff check ipg/ tests/"
format-check = "ruff format --check ipg/ tests/"
check = ["lint", "format-check"]

# Fix commands
lint-fix = "ruff check --fix ipg/ tests/"
format = "ruff format ipg/ tests/"
fix = ["format", "lint-fix"]

# Test commands
test = "pytest"
test-fast = "pytest -x"
test-verbose = "pytest -xvs"

[dependency-groups]
dev = [
    "faker>=40.4.0",
    "poethepoet>=0.41.0",
    "ruff>=0.15.2",
]
